+++ source/app/service-providers/stats/index.ts
@@ -14,12 +14,14 @@
  */
 
 import path from 'path'
+import { promises as fs } from 'fs'
 import { app, ipcMain } from 'electron'
 import ProviderContract from '../provider-contract'
 import type LogProvider from '../log'
 import PersistentDataContainer from '@common/modules/persistent-data-container'
 import broadcastIPCMessage from '@common/util/broadcast-ipc-message'
 import { today } from '@common/util/stats'
+import { disambiguateFile } from 'source/app/util/disambiguate-filename'
 
 // This is the data exposed publicly
 export interface Stats {
@@ -102,7 +104,10 @@ export default class StatsProvider extends ProviderContract {
     if (!await this.container.isInitialized()) {
       // Stats container is not yet initialized
       await this.container.init(this.stats)
-    } else {
+      return
+    }
+
+    try {
       const parsedData = await this.container.get()
       // We cannot safeAssign here, as we store everything in records, which
       // would make the function throw the keys away.
@@ -123,6 +128,15 @@ export default class StatsProvider extends ProviderContract {
           }
         }
       }
+    } catch (err: any) {
+      // If getting the data throws an error, we must not prevent Zettlr from
+      // starting. In this case, we'll init the container with our default data
+      // and move the malformed file. See #6127 for context.
+      const newPath = await disambiguateFile(this.statsFile)
+      await fs.rename(this.statsFile, newPath)
+      this._logger.error(`[StatsProvider] Could not read stats file: ${err.message}. Resetting stats to default and moving malformed file to backup: ${path.basename(newPath)}. Please check the file for what caused the error and possibly report this as a potential bug.`, err)
+      // Re-init with the correct data
+      await this.container.init(this.stats)
     }
 
     this._recompute()